test: move --hot stdin listener test to its own file
hot.test.ts has a pre-existing timing-sensitive test that fails on
debug builds; keep the stdin-listener coverage in a dedicated file so
it can be run in isolation. Fixture is inlined via tempDir().
